Move Native App entities to single modules
