--- src/snowflake/cli/_plugins/nativeapp/application_package_entity_model.py
@@ -1,94 +1,0 @@
-# Copyright (c) 2024 Snowflake Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from __future__ import annotations
-
-from typing import List, Literal, Optional, Union
-
-from pydantic import Field, field_validator
-from snowflake.cli.api.project.schemas.entities.common import (
-    EntityModelBase,
-    Identifier,
-)
-from snowflake.cli.api.project.schemas.updatable_model import (
-    DiscriminatorField,
-    IdentifierField,
-)
-from snowflake.cli.api.project.schemas.v1.native_app.package import DistributionOptions
-from snowflake.cli.api.project.schemas.v1.native_app.path_mapping import PathMapping
-from snowflake.cli.api.project.util import append_test_resource_suffix
-
-
-class ApplicationPackageEntityModel(EntityModelBase):
-    type: Literal["application package"] = DiscriminatorField()  # noqa: A003
-    artifacts: List[Union[PathMapping, str]] = Field(
-        title="List of paths or file source/destination pairs to add to the deploy root",
-    )
-    bundle_root: Optional[str] = Field(
-        title="Folder at the root of your project where artifacts necessary to perform the bundle step are stored.",
-        default="output/bundle/",
-    )
-    deploy_root: Optional[str] = Field(
-        title="Folder at the root of your project where the build step copies the artifacts",
-        default="output/deploy/",
-    )
-    generated_root: Optional[str] = Field(
-        title="Subdirectory of the deploy root where files generated by the Snowflake CLI will be written.",
-        default="__generated/",
-    )
-    stage: Optional[str] = IdentifierField(
-        title="Identifier of the stage that stores the application artifacts.",
-        default="app_src.stage",
-    )
-    scratch_stage: Optional[str] = IdentifierField(
-        title="Identifier of the stage that stores temporary scratch data used by the Snowflake CLI.",
-        default="app_src.stage_snowflake_cli_scratch",
-    )
-    distribution: Optional[DistributionOptions] = Field(
-        title="Distribution of the application package created by the Snowflake CLI",
-        default="internal",
-    )
-    manifest: str = Field(
-        title="Path to manifest.yml",
-    )
-
-    @field_validator("identifier")
-    @classmethod
-    def append_test_resource_suffix_to_identifier(
-        cls, input_value: Identifier | str
-    ) -> Identifier | str:
-        identifier = (
-            input_value.name if isinstance(input_value, Identifier) else input_value
-        )
-        with_suffix = append_test_resource_suffix(identifier)
-        if isinstance(input_value, Identifier):
-            return input_value.model_copy(update=dict(name=with_suffix))
-        return with_suffix
-
-    @field_validator("artifacts")
-    @classmethod
-    def transform_artifacts(
-        cls, orig_artifacts: List[Union[PathMapping, str]]
-    ) -> List[PathMapping]:
-        transformed_artifacts = []
-        if orig_artifacts is None:
-            return transformed_artifacts
-
-        for artifact in orig_artifacts:
-            if isinstance(artifact, PathMapping):
-                transformed_artifacts.append(artifact)
-            else:
-                transformed_artifacts.append(PathMapping(src=artifact))
-
-        return transformed_artifacts
