+++ src/snowflake/cli/_plugins/nativeapp/run_processor.py
@@ -18,13 +18,11 @@ from pathlib import Path
 from typing import Optional
 
 import typer
-from snowflake.cli._plugins.nativeapp.application_entity import (
-    ApplicationEntity,
-)
-from snowflake.cli._plugins.nativeapp.application_package_entity import (
+from snowflake.cli._plugins.nativeapp.artifacts import BundleMap
+from snowflake.cli._plugins.nativeapp.entities.application import ApplicationEntity
+from snowflake.cli._plugins.nativeapp.entities.application_package import (
     ApplicationPackageEntity,
 )
-from snowflake.cli._plugins.nativeapp.artifacts import BundleMap
 from snowflake.cli._plugins.nativeapp.manager import (
     NativeAppCommandProcessor,
     NativeAppManager,
