+++ src/snowflake/cli/_plugins/nativeapp/version/version_processor.py
@@ -17,7 +17,7 @@ from __future__ import annotations
 from pathlib import Path
 from typing import Dict, Optional
 
-from snowflake.cli._plugins.nativeapp.application_package_entity import (
+from snowflake.cli._plugins.nativeapp.entities.application_package import (
     ApplicationPackageEntity,
 )
 from snowflake.cli._plugins.nativeapp.manager import (
