+++ tests/nativeapp/utils.py
@@ -33,7 +33,7 @@ VERSION_MODULE = "snowflake.cli._plugins.nativeapp.version.version_processor"
 ENTITIES_COMMON_MODULE = "snowflake.cli.api.entities.common"
 ENTITIES_UTILS_MODULE = "snowflake.cli.api.entities.utils"
 APPLICATION_PACKAGE_ENTITY_MODULE = (
-    "snowflake.cli._plugins.nativeapp.application_package_entity"
+    "snowflake.cli._plugins.nativeapp.entities.application_package"
 )
 
 CLI_GLOBAL_TEMPLATE_CONTEXT = (
@@ -73,7 +73,7 @@ RUN_PROCESSOR_APP_POST_DEPLOY_HOOKS = f"{RUN_PROCESSOR}.app_post_deploy_hooks"
 
 FIND_VERSION_FROM_MANIFEST = f"{VERSION_MODULE}.find_version_info_in_manifest_file"
 
-APP_ENTITY_MODULE = "snowflake.cli._plugins.nativeapp.application_entity"
+APP_ENTITY_MODULE = "snowflake.cli._plugins.nativeapp.entities.application"
 APP_ENTITY = f"{APP_ENTITY_MODULE}.ApplicationEntity"
 APP_ENTITY_GET_EXISTING_APP_INFO = f"{APP_ENTITY}.get_existing_app_info"
 APP_ENTITY_DROP_GENERIC_OBJECT = f"{APP_ENTITY_MODULE}.drop_generic_object"
@@ -81,7 +81,7 @@ APP_ENTITY_GET_OBJECTS_OWNED_BY_APPLICATION = (
     f"{APP_ENTITY}.get_objects_owned_by_application"
 )
 
-APP_PACKAGE_ENTITY = "snowflake.cli._plugins.nativeapp.application_package_entity.ApplicationPackageEntity"
+APP_PACKAGE_ENTITY = "snowflake.cli._plugins.nativeapp.entities.application_package.ApplicationPackageEntity"
 APP_PACKAGE_ENTITY_DISTRIBUTION_IN_SF = (
     f"{APP_PACKAGE_ENTITY}.get_app_pkg_distribution_in_snowflake"
 )
